Let JSpinner and JComboBox (editable) draw their own background
The JTextField and JFormattedTextField now support a client property (joxy.isEditor) that switches off their background painting. That property is being set by JSpinner and the editable JComboBox on their JTextFields, to prevent that they draw their background, and JComboBox and JTextField draw their own background now.
Java by default has support for this, but we cannot use it, because we don´t use Borders for our... borders.