Hover and focus animation for JLists
This is something different than the hover decoration for individual elements. Also fixed a bug that caused the hover decoration to be incorrect for multicolumn JLists, so file choosers should work correctly now. There is only one remaining problem: the hover and focus animations, and the JList's borders, scroll with the JList contents if it is inside a JScrollPane. Some advanced hacks would probably be needed there.