Fix for bug <https://sourceforge.net/p/joxy/discussion/general/thread/e83a9187/>
It appears that jEdit (and many other applications, such as JOSM, as well) first creates a JFrame, then sets the LAF (without using SwingUtilities.updateComponentTreeUI(frame)). This breaks Joxy if the application switches from Joxy to another LAF, because the JoxyRootPaneUI delegate will stay on the GUI, relying on certain defaults not available anymore.

This fix checks, every time the JRootPane is painted, whether Joxy is still the installed LAF and if not, it calls SwingUtilities.updateComponentTreeUI(frame) itself. This is perhaps a bit inefficient, since the check is done on every paint. It could be improved by catching exceptions instead and calling the method there, but that of course is less robust, since we will have to track very carefully which defaults are used where.