Start implementing caching for each painter
This means that painters are no longer just static methods in painter classes, but painters become objects that are able to cache what they painted for a certain size. For this to be effective, every delegate that uses a painter must maintain its own instance of the painter.

A new abstract super class Painter is used to handle the caching. The painter classes then only need to override paintObject.

For now, only the ButtonSlabPainter has been converted.