Severe performance bug fixed in the custom borders
JoxyPanelUI, on drawing of the panel, used to check whether a Java border was added to the component, and if so, swap it by a Joxy border. This setBorder call will automatically schedule a repaint.

This check was done by using the following:

    if (b instanceof BevelBorder) { ... }

and so on for the other border types.

But that is incorrect: if b already was a Joxy border, this would also be true (since Joxy borders extend their Java equivalents). So, every border would be painted continuously, which meant that every application using borders completely started using 100% CPU.

Fixed this now by adding additionally

    !(b instanceof JoxyBevelBorder)