Bug 10: paint button background colour based on their position
In Oxygen, the background colour of buttons is dependent on their y coordinate in the window. This way, they fit in nicely with the background gradient.

Joxy does this too now. It is not perfect yet, and the exact colours in the gradient are a guess. The base colour of the button´s background is determined by code that we translated from the original Oxygen code, so that is probably correct.