Add switching of modes to the JColorChooser
One can now choose between Red, Green and Blue mode. Furthermore the GUI for changing the colour numerically has been created.

The switching has been implemented in DiagramComponent by adding an enum DiagramComponentMode, that contains values for the three (should become six in the future) possible modes; furthermore, a ColorModeCalculator is provided for all of these modes. Such a calculator is able to convert colours to positions in the colour square and the other way round. So, to add new modes, the only thing to do is (except for modifying the GUI of course) to add a DiagramComponentMode to the enum, along with its ColorModeCalculator. Then the other things are handled automatically.