Rework the JCheckBox and JRadioButton support
Merge the two (nearly identical) delegates for JCheckBoxes and JRadioButtons in one class, JCheckRadioUI. This delegate is used for both, and it detects automatically which one to draw. This also required changes to the painters (merging ButtonSlabPainter and RoundButtonSlabPainter).

At the moment there is still an issue with the focus/hover highlight of JRadioButtons because the painter for this also needs to be reworked; this will be done soon.