Added a class that can be used to find some KDE paths, for example the directory where configuration files are stored, or icons.
This works by directly calling `kde4-config`. I have written a test that checks that indeed, valid paths are returned. Note that when `kde4-config` is not available, an error is issued (via the `Output` class) and an empty list is returned. The same holds when some unexpected error occurs.

Using this class in `Utils` should make the code more robust. Way more robust.